requirements.txt
# Dependências principais
langchain==0.1.12
langchain-community==0.0.28
supabase==2.3.4
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# Dependências para integração com IA
ollama==0.1.6
chromadb==0.4.24

# Dependências para processamento de dados
pandas==2.2.1
numpy==1.26.4

# Dependências para testes
pytest==8.0.2
pytest-asyncio==0.23.5

# Dependências para desenvolvimento
black==24.2.0
isort==5.13.2
flake8==7.0.0 

# Dependências para o FastAPI e o PostgreSQL do Supabase
psycopg2-binary==2.9.9

